Make optional integrations package extensions
