+++ Project.toml
@@ -6,18 +6,24 @@ version = "0.2.0"
 [deps]
 AxisArrays = "39de3d68-74b9-583c-8d2d-e117c070f3a9"
 DimensionfulAngles = "2d4b8d7a-02d9-40f9-9abe-9c695b77de0d"
-Integrals = "de52edbc-65ea-441a-8357-d3a637375a31"
-Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
 Roots = "f2b01f46-fcfa-551c-844a-d8ac1e96c665"
 Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 
+[weakdeps]
+Integrals = "de52edbc-65ea-441a-8357-d3a637375a31"
+Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+
+[extensions]
+WaveSpectraIntegralsExt = "Integrals"
+WaveSpectraMakieExt = "Makie"
+
 [compat]
 AxisArrays = "0.4"
 DimensionfulAngles = "0.2"
-Integrals = "5"
-Makie = "0.24"
+Integrals = ">= 0"
+Makie = "0.20, 0.21, 0.22, 0.23, 0.24"
 Plots = "1"
 PrettyTables = "3"
 Roots = "2, 3.0"
