+++ src/WaveSpectra.jl
@@ -13,11 +13,13 @@ using Unitful: Dimensions, FreeUnits, NoDims, Quantity, Units, dimension, gn as 
                m, s, ustrip, 𝐋, 𝐓
 using DimensionfulAngles: Dispersion, turnᵃ as τ, θ₀, 𝐀
 using DimensionfulAngles.DefaultSymbols: rad, °
-using Integrals: AbstractSampledIntegralAlgorithm, TrapezoidalRule, UniformWeights
 using AxisArrays: Axis, AxisArray, ClosedInterval, axisvalues, (..)
 using Plots: mm as plots_mm, text, @recipe
 using PrettyTables: HtmlTableStyle, pretty_table
-using Makie: Axis as MAxis, Colorbar, Figure, PolarAxis, contourf!, lines!, plot!
+
+abstract type AbstractSampledIntegralAlgorithm end
+abstract type UniformWeights end
+struct TrapezoidalRule <: AbstractSampledIntegralAlgorithm end
 
 # imports to overload
 import Base  # BroadcastStyle, copy, eltype, getindex, isapprox, setindex!, show, similar,
@@ -25,7 +27,6 @@ import Base  # BroadcastStyle, copy, eltype, getindex, isapprox, setindex!, show
 import Unitful: uconvert, unit
 import AxisArrays # axes # in the future, do `import AxisArrays: axes as AAaxes`
 const axes = Base.axes # name conflict will be fixed by AxisArrays in the future
-import Integrals: find_weights
 
 # export
 export DispersionRelations, Moments, ParametricSpectra, Shapes # modules
